feat(domain): 덱 초기화 시 중복된 카드 검증 로직 구현
덱 초기화 시 중복된 카드가 들어있을 경우를 방지 하기 위해서 중복 카드를 검증하는 로직을 추가했습니다.
